Make `copilot` disabled globally by default
This commit changes the default behavior of `copilot` to be disabled
globally by default. This is done by setting `g:copilot_enabled` to
`v:false` in `copilot.vim`. For fast editing, that `nvim` is great for,
copilot is simply a overkill and merely exhausts the CPU/API. To toggle
it on, press `<F12>`.

Furthermore, now all `json` files are disabled by default. This is to
avoid bloat.